+++ src/routes/threads/index.ts
@@ -1,13 +1,17 @@
 import { FastifyInstance, FastifyPluginOptions } from "fastify";
-import search from "./search";
+import create from "./create";
 import threads from "./threads";
+import search from "./search";
+import thread from "./thread";
 
-export default function (
+export default (
     fastify: FastifyInstance,
     _opts: FastifyPluginOptions,
-    done: (err?: Error) => void
-) {
+    done: (e?: Error) => void
+) => {
     fastify.register(threads);
     fastify.register(search);
+    fastify.register(create);
+    fastify.register(thread, { prefix: "/:id" });
     done();
-}
+};