+++ src/routes/users/index.ts
@@ -5,12 +5,14 @@ import verify from "./verify";
 import resend from "./resend";
 import reset from "./reset";
 import forgot from "./forgot";
+import user from "./user";
 
 export default (
     fastify: FastifyInstance,
     _opts: FastifyPluginOptions,
     done: (e?: Error) => void
 ) => {
+    fastify.register(user, { prefix: "/:id" });
     fastify.register(register);
     fastify.register(login);
     fastify.register(verify);