+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## v5.1.0
+
+- remove `POST /me/rename`
+- move `POST /me/avatar` to `PUT /users/{id}/avatar`
+- return 404 error if avatar not found
+
+## v5.0.0
+
+- redesign rest api (reduce redundancy and more rest)
+  - merge /user to /users
+  - merge /thread to /threads
+  - move /category to /categories
+  - no longer use `POST /create`, instead `POST /`
+
 ## v4.8.0
 
 - move pin and unpin to `PUT` and `DELETE` `/thread/{id}/pin`